+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -11,5 +11,5 @@ public interface FilmStorage {
 
     Film updateFilm(Film film);
 
-    Film getFilmById(int id);
+    Film getFilmById(long id);
 }
